feat: graceful K8s termination
When deleting the pod, we receive the SIGTERM signal and a default grace
period of 30s after which we're force killed with SIGKILL.
This change adds a handler for SIGTERM allows us to stop the csi server
gracefully (if possible) and quickly terminate without having to wait.

Added SIGTERM handling as well, alongside with logs for information.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>